+++ tests/code/steane/test_steane_implement.py
@@ -2,7 +2,7 @@ from guppylang import guppy
 from guppylang.emulator import EmulatorBuilder
 from guppylang.std.platform import result
 
-from guppyft.code.steane.encoder_spec import SteaneSpec
+from guppyft.code.steane.encoder_spec import SteaneBuilder
 from guppyft.logical.steane import Qubit, cx, h, measure_z, x, z
 
 
@@ -14,8 +14,8 @@ def test_qalloc_measure() -> None:
         result("res", measure_z(q).decode())
 
     pkg = main.compile()
-    phys_pkg = SteaneSpec(n_blocks=1).implement_ops(pkg)
-    res = EmulatorBuilder().build(phys_pkg, n_qubits=7).run().collated_shots()
+    phys_pkg = SteaneBuilder().build(n_blocks=1).implement_ops(pkg)
+    res = EmulatorBuilder().build(phys_pkg, n_qubits=8).run().collated_shots()
 
     assert res == [{"res": [0]}]
 
@@ -29,8 +29,8 @@ def test_x() -> None:
         result("res", measure_z(q).decode())
 
     pkg = main.compile()
-    phys_pkg = SteaneSpec(n_blocks=1).implement_ops(pkg)
-    res = EmulatorBuilder().build(phys_pkg, n_qubits=7).run().collated_shots()
+    phys_pkg = SteaneBuilder().build(n_blocks=1).implement_ops(pkg)
+    res = EmulatorBuilder().build(phys_pkg, n_qubits=8).run().collated_shots()
 
     assert res == [{"res": [1]}]
 
@@ -46,8 +46,8 @@ def test_h_z() -> None:
         result("res", measure_z(q).decode())
 
     pkg = main.compile()
-    phys_pkg = SteaneSpec(n_blocks=1).implement_ops(pkg)
-    res = EmulatorBuilder().build(phys_pkg, n_qubits=7).run().collated_shots()
+    phys_pkg = SteaneBuilder().build(n_blocks=1).implement_ops(pkg)
+    res = EmulatorBuilder().build(phys_pkg, n_qubits=8).run().collated_shots()
 
     assert res == [{"res": [1]}]
 
@@ -63,7 +63,7 @@ def test_cx() -> None:
         result("q1", measure_z(q1).decode())
 
     pkg = main.compile()
-    phys_pkg = SteaneSpec(n_blocks=2).implement_ops(pkg)
-    res = EmulatorBuilder().build(phys_pkg, n_qubits=14).run().collated_shots()
+    phys_pkg = SteaneBuilder().build(n_blocks=2).implement_ops(pkg)
+    res = EmulatorBuilder().build(phys_pkg, n_qubits=18).run().collated_shots()
 
     assert res == [{"q0": [1], "q1": [1]}]